+++ scripts/west_commands/export_app/Changelog.md
@@ -14,8 +14,8 @@
 ## [25.09.00]
 
 - New Features
-  - Support new option "--bf", this will copy board files to the output directory.
-  - User can set "freestanding_copied_folders" in example.yml => example => contents to specify default board copy folders.
+  - Support new option `--bf`, this will copy board files to the output directory.
+  - User can set `freestanding_copied_folders` in example.yml => example => contents to append board copy folders.
   - Support set output_dir and clean_output_dir in west config file.
 
 - Improvements